+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1291,5 +1291,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-06T08:24:22.146Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-06T11:41:10.403Z")
     r"""The date and time the resource was last updated"""