+++ packages/orq-rc/src/orq_ai_sdk/models/getagentop.py
@@ -303,7 +303,7 @@ class GetAgentKnowledgeBases(BaseModel):
     configuration: GetAgentKnowledgeBaseConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K6W9YKAYEZ2ARCY855M2P4V8"
+    id: Optional[str] = "01K6WN6YEVS1ETWM3DZ7XT8GA3"
     r"""The id of the resource"""