+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6W9YKETCW9W6SD46Q47GQFW"
+        "01K6WN6YKG5TYD42DSPM4J3QB1"
     )
     r"""The id of the resource"""