+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -670,5 +670,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-06T08:24:22.146Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-06T11:41:10.403Z")
     r"""The date and time the resource was last updated"""