+++ packages/orq-rc/RELEASES.md
@@ -3149,3 +3149,13 @@ Based on:
 - [python v3.14.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.9] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.9 - packages/orq-rc
+
+## 2025-10-06 11:42:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.634.2 (2.721.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.10] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.10] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.10 - packages/orq-rc